test: move http tests into separate file
